+++ src/main/java/exceptions/DukeException.java
@@ -1,4 +1,4 @@
-package utils.exceptions;
+package exceptions;
 
 public class DukeException extends Exception {
     public static final String ERROR_UNEXPECTED = "Error: Something unexpected happened. Please submit a bug report.";
