+++ src/main/java/task/Task.java
@@ -1,4 +1,4 @@
-package utils.task;
+package task;
 
 /**
  * The Task class encapsulates a task.
