+++ src/main/java/task/TaskList.java
@@ -1,6 +1,6 @@
-package utils.task;
+package task;
 
-import utils.exceptions.DukeException;
+import exceptions.DukeException;
 
 import java.util.ArrayList;
